refactor:[R6-trivial] Renomeia pacote injeção de dependências
